Auth provider to user ccontroller’s methods

+++ resources/views/user/index.blade.php
@@ -61,15 +61,19 @@
 			                        			<td>{{ date('Y-m-d, H:m', strtotime($user->updated_at)) }}</td>
 			                        			{!! $user->deleted_at ? '<td class="red">Deleted</td>' : '<td class="blue">Aktif</td>' !!}
 			                        			<td>
-			                        				@if(!$user->deleted_at)
+			                        				@if(!$user->deleted_at && Gate::check('edit_u'))
 			                        					<a class="btn btn-sm btn-primary" href="{{ url('user').'/'.$user->id.'/edit' }}"><i class="fa fa-edit"></i> Edit</a>
 		                        					@endif
-		                        					@if(Auth::user()->id != $user->id && !$user->deleted_at)
+		                        					@if(Auth::user()->id != $user->id && !$user->deleted_at && Gate::check('sdelete_u'))
 				                        				<a class="btn btn-sm btn-danger" href="#" onclick="deleteUser({{ $user->id }}, false)"><i class="fa fa-times"></i> Hapus</a>
 				                        			@endif
 				                        			@if($user->deleted_at)
-				                        				<a class="btn btn-sm btn-warning" href="#" onclick="restoreUser({{ $user->id }})"><i class="fa fa-backward"></i> Restore</a>
-				                        				<a class="btn btn-sm btn-danger" href="#" onclick="deleteUser({{ $user->id }}, true)"><i class="fa fa-times"></i> Hapus permanen</a>
+				                        				@can('restore_u')
+				                        				    <a class="btn btn-sm btn-warning" href="#" onclick="restoreUser({{ $user->id }})"><i class="fa fa-backward"></i> Restore</a>
+				                        				@endcan
+				                        				@can('pdelete_u')
+				                        				    <a class="btn btn-sm btn-danger" href="#" onclick="deleteUser({{ $user->id }}, true)"><i class="fa fa-times"></i> Hapus permanen</a>
+				                        				@endcan
 			                        				@endif
 		                        				</td>
 			                        		</tr>
